+++ ChangeLog.md
@@ -1,4 +1,6 @@
 # Prophets of the Labyrinth Change Log
+## Prophets of the Labyrinth v0.14.0:
+- Combined event rooms "Elemental Research" and "HP Redistribution" into "Imp Contract Faire"
 ## Prophets of the Labyrinth v0.13.0:
 - New Gear: Refreshing Breeze, Shattering Sabotage Kit, Harmful Corrosion, Shattering Corrosion, Wolf Ring, Surpassing Wolf Ring, Swift Wolf Ring, Harmful Poison Torrent, Ice Bolt, Scarf, Hearty Scarf, Chainmail
 - Reworked Thick Cloak to Accurate Cloak